Move tests to be able to run from openapi
With python -m unittest
